To test signals

+++ sw/applications/kernel_test/kernels/strsearch/strsearch.c
@@ -147,14 +147,14 @@ void config()
 	}
     PRINTDBG("\n");
 
-	cgra_input[1][0] = i_patlen_cgra;
-	cgra_input[1][1] = i_skip2_cgra;
+	cgra_input[1][0] = (uint32_t)i_patlen_cgra;
+	cgra_input[1][1] = (uint32_t)i_skip2_cgra;
 
-	cgra_input[2][0] = lowervec;
-	cgra_input[2][1] = lowerc(i_pattern_cgra[i_patlen_cgra - 1]);
+	cgra_input[2][0] = (uint32_t)lowervec;
+	cgra_input[2][1] = (uint32_t)lowerc(i_pattern_cgra[i_patlen_cgra - 1]);
 
-	cgra_input[3][0] = i_pattern_cgra;
-	cgra_input[3][1] = i_patlen_cgra -1;
+	cgra_input[3][0] = (uint32_t)i_pattern_cgra;
+	cgra_input[3][1] = (uint32_t)(i_patlen_cgra -1);
     
     PRINTDBG("[1][0] = %08x\n", i_patlen_cgra);
     PRINTDBG("[1][1] = %08x\n", i_skip2_cgra);
